Moved EndPointType to its own file.
